# src/main/resources/openapi/pms-api.yaml
openapi: 3.0.3
info:
  title: PMS 项目管理 API
  version: 1.0.0
  description: 单企业私有化部署的项目、组织和权限管理接口。
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /api
...
paths:
...
  /projects/page:
    post:
      tags: [项目]
      summary: 项目分页查询
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                keyword: { type: string }
                status: { type: integer }
                view: { type: string, enum: [MINE, PORTFOLIO, ALL] }
                orgUnitId: { type: integer, format: int64 }
                projectManagerId: { type: integer, format: int64 }
                projectLevel: { type: integer, enum: [0, 1, 2, 3] }
                attention: { type: string, enum: [ACTIVE, OVERDUE, NO_MANAGER, STALE_NODE] }
                currentNodeKey: { type: string }
                currPage: { type: integer, minimum: 1, default: 1 }
                pageSize: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /projects:
    post:
      tags: [项目]
      summary: 创建项目
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                priority: { type: integer, minimum: 0, maximum: 3, description: 低、中、高、紧急 }
                projectLevel: { type: integer, enum: [0, 1, 2, 3], default: 0, description: 0常规(C)、1重要(B)、2关键(A)、3战略(S) }
                orgUnitId: { type: integer, format: int64 }
                startDate: { type: string, format: date }
                endDate: { type: string, format: date }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /projects/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [项目]
      summary: 项目详情
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [项目]
      summary: 更新项目
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                priority: { type: integer, minimum: 0, maximum: 3, description: 低、中、高、紧急 }
                projectLevel: { type: integer, enum: [0, 1, 2, 3], description: 0常规(C)、1重要(B)、2关键(A)、3战略(S) }
                orgUnitId: { type: integer, format: int64 }
                projectManagerId: { type: integer, format: int64, nullable: true }
                memberIds: { type: array, items: { type: integer, format: int64 } }
                expectedMemberIds: { type: array, items: { type: integer, format: int64 }, description: 客户端上次看到的成员集合，用于避免覆盖并发自动加入的成员 }
                followerIds: { type: array, items: { type: integer, format: int64 } }
                startDate: { type: string, format: date }
                endDate: { type: string, format: date }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [项目]
      summary: 删除项目
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /projects/stats:
    get:
      tags: [项目]
      summary: 项目统计
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /projects/summary:
    post:
      tags: [项目]
      summary: 项目列表摘要
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                keyword: { type: string }
                status: { type: integer }
                view: { type: string, enum: [MINE, PORTFOLIO, ALL] }
                orgUnitId: { type: integer, format: int64 }
                projectManagerId: { type: integer, format: int64 }
                projectLevel: { type: integer, enum: [0, 1, 2, 3] }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /projects/{id}/terminate:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      tags: [项目]
      summary: 终止项目
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /projects/{id}/restore:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      tags: [项目]
      summary: 恢复已终止项目
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /projects/{projectId}/nodes:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [任务与节点]
      summary: 项目节点列表
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /projects/{projectId}/nodes/{nodeId}/complete:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/NodeId'
    post:
      tags: [任务与节点]
      summary: 完成节点
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '409': { $ref: '#/components/responses/Conflict' }
  /projects/{projectId}/nodes/{nodeId}/rollback:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/NodeId'
    post:
      tags: [任务与节点]
      summary: 回滚节点
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '409': { $ref: '#/components/responses/Conflict' }
  /projects/{projectId}/nodes/{nodeId}/owner:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/NodeId'
    put:
      tags: [任务与节点]
      summary: 设置节点负责人
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /projects/{projectId}/nodes/{nodeId}/schedule:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/NodeId'
    put:
      tags: [任务与节点]
      summary: 更新节点排期
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { $ref: '#/components/responses/Success' }
...
  /projects/{projectId}/members:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [项目]
      summary: 项目成员列表
      responses:
        '200': { $ref: '#/components/responses/Success' }
    post:
      tags: [项目]
      summary: 添加项目成员
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /projects/{projectId}/members/{memberId}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - name: memberId
        in: path
        required: true
        schema: { type: integer, format: int64 }
    delete:
      tags: [项目]
      summary: 移除项目成员
      responses:
        '200': { $ref: '#/components/responses/Success' }
...
  /projects/{projectId}/comments:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [项目]
      summary: 项目动态列表；传入 taskId 时仅返回该任务评论
      parameters:
        - name: taskId
          in: query
          schema: { type: integer, format: int64 }
      responses:
        '200': { $ref: '#/components/responses/Success' }
    post:
      tags: [项目]
      summary: 发布项目动态
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /comments/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    delete:
      tags: [项目]
      summary: 删除动态
      responses:
        '200': { $ref: '#/components/responses/Success' }
...
  /admin/import/preview/organizations:
    post:
      tags: [导入与审计]
      summary: 预览组织 Excel/CSV 导入
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '422': { $ref: '#/components/responses/BusinessError' }
  /admin/import/preview/users:
    post:
      tags: [导入与审计]
      summary: 预览员工 Excel/CSV 导入
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '422': { $ref: '#/components/responses/BusinessError' }
  /admin/import/{jobId}/commit:
    parameters:
      - name: jobId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [导入与审计]
      summary: 提交导入任务
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '409': { $ref: '#/components/responses/Conflict' }
  /admin/import/{jobId}/errors.csv:
    parameters:
      - name: jobId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [导入与审计]
      summary: 下载导入错误报告
      responses:
        '200': { description: CSV 错误报告 }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: { type: integer, format: int64 }
    NodeId:
      name: nodeId
      in: path
      required: true
      schema: { type: integer, format: int64 }
  schemas:
...
    ApiResponse:
      type: object
      properties:
        code: { type: integer, example: 0 }
        msg: { type: string }
        data: {}
        requestId: { type: string }
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          required: [code, msg, requestId]
  responses:
    Success:
      description: 请求成功
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiResponse' }
    Unauthorized:
      description: 未认证或会话已失效
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: 无权访问
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: 数据冲突
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    BusinessError:
      description: 业务校验失败
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
security:
  - bearerAuth: []